requirements.txt
mkdocs-material~=9.6.14
mkdocs-publisher~=1.4.7
mkdocs-exclude~=1.0.2
mkdocs-awesome-nav~=3.1.2
mkdocs-obsidian-bridge~=1.2.0
mkdocs-same-dir~=0.1.3
